+++ mycodo/mycodo_flask/utils/utils_method.py
@@ -111,7 +111,7 @@ def method_create(form_create_method):
     if dep_unmet:
         list_unmet_deps = []
         for each_dep in dep_unmet:
-            list_unmet_deps.append(each_dep[0])
+            list_unmet_deps.append(each_dep[3])
         error.append(
             "The {dev} device you're trying to add has unmet dependencies: "
             "{dep}".format(dev=form_create_method.method_type.data,