+++ mycodo/outputs/on_off_kp303.py
@@ -51,7 +51,7 @@ channels_dict = {
 # Output information
 OUTPUT_INFORMATION = {
     'output_name_unique': 'kp303',
-    'output_name': f"{lazy_gettext('On/Off')}: KP303 Kasa Smart WiFi Power Strip",
+    'output_name': f"{lazy_gettext('On/Off')}: KP303 Kasa Smart WiFi Power Strip (python-kasa 0.4.0.dev4)",
     'output_manufacturer': 'TP-Link',
     'input_library': 'python-kasa',
     'measurements_dict': measurements_dict,