+++ mycodo/scripts/generate_manual_widgets.py
@@ -77,9 +77,6 @@ if __name__ == "__main__":
                                     out_file.write("[{0}](https://pypi.org/project/{0})".format(each_dep[2].split("==")[0]))
                                 else:
                                     out_file.write("[{0}](https://pypi.org/project/{0})".format(each_dep[2]))
-                            elif each_dep[0] == "pip-git":
-                                url = re.search('git://(.*).git', each_dep[2])
-                                out_file.write("[{name}](https://{url})".format(name=each_dep[1], url=url.group(1)))
                             elif each_dep[0] == "bash-commands":
                                 list_files = []
                                 for each_path in each_dep[1]: